Classic v1.31

+++ include/mylib/gfx/texture.hpp
@@ -1,12 +1,13 @@
 #ifndef TEXTURE_HPP
 #define TEXTURE_HPP
 
+#include <cstdint>
 
 extern void loadTexture(const char* fileName);
-extern void createTexture(const char* filePath, unsigned int& texture);
-inline constexpr unsigned int atlasWidth {2}, atlasHeight {2};
+extern void createTexture(const char* filePath, uint32_t& texture);
+inline constexpr uint16_t atlasWidth {2}, atlasHeight {2};
 inline constexpr float xPos {1.0f / atlasWidth}, yPos {1.0f / atlasHeight};
-inline unsigned int textureID {};
+inline uint32_t textureID {};
 
 
 #endif // TEXTURE_HPP
