+++ src/main.cpp
@@ -3,8 +3,8 @@
 
 int main()
 {
-    constexpr unsigned int windowWidth {400}, windowHeight {400};
-    auto  minecraft = std::make_unique<Game>(windowWidth, windowHeight);
+    constexpr uint16_t windowWidth {400}, windowHeight {400};
+    auto minecraft = std::make_unique<Game>(windowWidth, windowHeight);
     minecraft->runGame();
     return 0;
 }
